Migrate onondaga (Election Book PDF via natural_pdf) — completes election_book family
